+++ src/AngleSharp.Xml/Dom/Internal/XmlElement.cs
@@ -1,6 +1,8 @@
 namespace AngleSharp.Xml.Dom
 {
     using AngleSharp.Dom;
+    using AngleSharp.Text;
+    using AngleSharp.Xml.Parser;
     using System;
 
     /// <summary>
@@ -10,8 +12,8 @@ namespace AngleSharp.Xml.Dom
     {        
         #region ctor
 
-        public XmlElement(Document owner, String name, String prefix = null)
-            : base(owner, name, prefix, null)
+        public XmlElement(Document owner, String name, String prefix = null, String namespaceUri = null, NodeFlags flags = NodeFlags.None)
+            : base(owner, name, prefix, namespaceUri, flags)
         {
         }
 
@@ -29,6 +31,21 @@ namespace AngleSharp.Xml.Dom
 
         #region Methods
 
+        /// <inheritdoc />
+        public override IElement ParseSubtree(String html)
+        {
+            var context = ((IElement)this).Owner?.Context;
+            var source = new TextSource(html);
+            var document = new XmlDocument(context, source);
+            var parser = new XmlDomBuilder(document);
+            var options = new XmlParserOptions
+            {
+                IsSuppressingErrors = true,
+            };
+            return parser.ParseFragment(options, this).DocumentElement;
+        }
+
+        /// <inheritdoc />
         public override Node Clone(Document owner, Boolean deep)
         {
             var node = new XmlElement(owner, LocalName, Prefix);