+++ src/AngleSharp.Xml/Dtd/Parser/DtdParser.cs
@@ -53,17 +53,14 @@ namespace AngleSharp.Xml.Dtd.Parser
         /// </summary>
         public Boolean IsInternal
         {
-            get { return _tokenizer.IsExternal; }
-            set { _tokenizer.IsExternal = !value; }
+            get => _tokenizer.IsExternal;
+            set => _tokenizer.IsExternal = !value;
         }
 
         /// <summary>
         /// Gets the DtdContainer, i.e. the result of the computation.
         /// </summary>
-        public DtdContainer Result
-        {
-            get { return _result; }
-        }
+        public DtdContainer Result => _result;
 
         #endregion