+++ src/AngleSharp.Xml/Parser/XmlDomBuilder.cs
@@ -119,6 +119,23 @@ namespace AngleSharp.Xml.Parser
             return _document;
         }
 
+        /// <summary>
+        /// Switches to the fragment algorithm with the specified context
+        /// element. Then parses the given source and creates the document.
+        /// </summary>
+        /// <param name="options">The options to use for parsing.</param>
+        /// <param name="context">
+        /// The context element where the algorithm is applied to.
+        /// </param>
+        public Document ParseFragment(XmlParserOptions options, Element context)
+        {
+            context = context ?? throw new ArgumentNullException(nameof(context));
+            var root = new XmlElement(_document, context.LocalName);
+            _document.AddNode(root);
+            _openElements.Add(root);
+            return Parse(options);
+        }
+
         #endregion
 
         #region States
@@ -254,7 +271,7 @@ namespace AngleSharp.Xml.Parser
                 case XmlTokenType.StartTag:
                 {
                     var tagToken = (XmlTagToken)token;
-                    var element = CreateElement(tagToken.Name);
+                    var element = CreateElement(tagToken.Name, tagToken.IsSelfClosing);
                     CurrentNode.AppendChild(element);
 
                     for (var i = 0; i < tagToken.Attributes.Count; i++)
@@ -385,15 +402,11 @@ namespace AngleSharp.Xml.Parser
 
         #region Helpers
 
-        private static Element CreateElement(Document document, String name, String prefix)
-        {
-            return new XmlElement(document, name, prefix);
-        }
-
-        private Element CreateElement(String name)
+        private Element CreateElement(String name, Boolean selfClosing)
         {
             var prefix = default(String);
             var colon = name.IndexOf(Symbols.Colon);
+            var flags = selfClosing ? NodeFlags.SelfClosing : NodeFlags.None;
 
             if (colon > 0 && colon < name.Length - 1)
             {
@@ -401,7 +414,7 @@ namespace AngleSharp.Xml.Parser
                 name = name.Substring(colon + 1);
             }
 
-            return _document.CreateElementFrom(name, prefix);
+            return _document.CreateElementFrom(name, prefix, flags);
         }
 
         private Attr CreateAttribute(String name, String value)