+++ CHANGELOG.md
@@ -1,3 +1,41 @@
+# 0.15.0
+
+Released on Thursday, April 22 2021.
+
+- Dropped support for .NET Standard 1.3 target
+- Fixed breaking parsing upon invalid syntax when suppressing errors (#14)
+
+# 0.14.0
+
+Released on Tuesday, March 31 2020.
+
+- Added more overloads for `ToXml` and `ToMarkup`
+- Allow transforming all empty elements to self closing tags
+- Serialize self-closing tags correctly (#11)
+- Updated the `IMarkupFormatter` implementations
+- Added shallow support for source link
+- Support more target frameworks
+
+# 0.13.0
+
+Released on Friday, September 6 2019.
+
+- Added `ToXml` and `ToMarkup` extensions for the formatting
+
+# 0.12.1
+
+Released on Wednesday, May 15 2019.
+
+- Binary version fix
+- Placed `WithXml` in the AngleSharp namespace (#4)
+
+# 0.12.0
+
+Released on Thursday, May 2 2019.
+
+- Reference latest AngleSharp
+- Improved `InnerHtml` / XML fragment parsing (#1)
+
 # 0.11.0
 
 Released on Tuesday, February 12 2019.